refactor: move SSRStore to a separate file and update imports
