+++ packages/gene-next/src/translations/index.ts
@@ -2,5 +2,5 @@ export * from './i18n';
 export * from './ioc/ioc';
 export * from './types';
 export * from './ssr';
-export {transChoiceHelper} from './utils/transChoiceHelper';
-export type {TransChoiceHelperFnType} from './utils/transChoiceHelper';
+export { transChoiceHelper } from './utils/transChoiceHelper';
+export type { TransChoiceHelperFnType } from './utils/transChoiceHelper';
