+++ packages/gene-next/src/translations/ssr/ssrFallback.ts
@@ -1,4 +1,4 @@
-import { SSRStore } from '../i18n';
+import { SSRStore } from './ssrStore';
 
 export function ssrFallback(key: string) {
   return SSRStore.current[key] || key;
