+++ packages/gene-next/src/translations/ssr/ssrStore.ts
@@ -1,0 +1,6 @@
+interface SSRStoreType {
+    current: Record<string, string>;
+  }
+  
+  export const SSRStore: SSRStoreType = { current: {} };
+
